Implement image upload for edit contact

+++ backend/routes/ContactRoutes.js
@@ -26,52 +26,64 @@ const upload = multer({
 router.post(
   "/add",
   upload.single("pfp"),
-  // validate(["name", "phone", "position", "organization", "email", "pfp", "crop"]),
+  validate(["name", "phone", "position", "organization", "email"]),
   (req, res) => {
-    // console.log(req.pfp);
+    if (req.file && req.file.path) {
+      const crop = JSON.parse(req.body.crop);
 
-    const crop = JSON.parse(req.body.crop);
+      let compressor = sharp(req.file.path).rotate();
+      if (crop.width && crop.height && crop.left && crop.top) {
+        compressor = compressor.extract(crop);
+      }
+      compressor = compressor.resize(400, 400).png({
+        compressionLevel: 9,
+        adaptiveFiltering: true,
+        force: true,
+      });
 
-    let compressor = sharp(req.file.path).rotate();
-    if (crop.width && crop.height && crop.left && crop.top) {
-      compressor = compressor.extract(crop);
-    }
-    compressor = compressor.resize(400, 400).png({
-      compressionLevel: 9,
-      adaptiveFiltering: true,
-      force: true,
-    });
+      Promise.all([
+        Contact.create({
+          name: req.body.name,
+          position: req.body.position,
+          organization: req.body.organization,
+          email: req.body.email,
+          phone: req.body.phone,
+        }),
+        uploadFileStream(
+          compressor,
+          `pfp/${req.file.filename}-${Date.now()}-${Math.round(Math.random() * 1e9)}`
+        ),
+      ])
+        .then(([contact, result]) => {
+          const old = contact.profilePicture;
 
-    Promise.all([
+          Object.assign(contact, {
+            profilePicture: result.key,
+            profilePictureModified: new Date(),
+          });
+
+          return Promise.all([
+            contact.toJSON(),
+            contact.save(),
+            old ? deleteFileAWS(old) : null,
+            fs.unlink(req.file.path),
+          ]);
+        })
+        .then(([contact]) => res.json({ success: true, contact }))
+        .catch(errorHandler(res));
+    } else {
       Contact.create({
         name: req.body.name,
         position: req.body.position,
         organization: req.body.organization,
         email: req.body.email,
         phone: req.body.phone,
-      }),
-      uploadFileStream(
-        compressor,
-        `pfp/${req.file.filename}-${Date.now()}-${Math.round(Math.random() * 1e9)}`
-      ),
-    ])
-      .then(([contact, result]) => {
-        const old = contact.profilePicture;
-
-        Object.assign(contact, {
-          profilePicture: result.key,
-          profilePictureModified: new Date(),
-        });
-
-        return Promise.all([
-          contact.toJSON(),
-          contact.save(),
-          old ? deleteFileAWS(old) : null,
-          fs.unlink(req.file.path),
-        ]);
+        profilePicture: null,
+        profilePictureModified: Date.now(),
       })
-      .then(([contact]) => res.json({ success: true, contact }))
-      .catch(errorHandler(res));
+        .then((contact) => res.json({ success: true, contact }))
+        .catch(errorHandler(res));
+    }
   }
 );
 
@@ -95,20 +107,64 @@ router.delete("/delete/:id", idParamValidator(), (req, res) => {
     .catch(errorHandler(res));
 });
 
-router.patch("/edit/:id", idParamValidator(), (req, res) => {
-  Contact.findById(req.params.id)
-    .then((contact) => {
-      Object.assign(contact, {
-        name: req.body.updated_name ? req.body.updated_name : contact.name,
-        position: req.body.updated_position ? req.body.updated_position : contact.position,
-        organization: req.body.updated_org ? req.body.updated_org : contact.organization,
-        email: req.body.updated_email ? req.body.updated_email : contact.email,
-        phone: req.body.updated_phone ? req.body.updated_phone : contact.phone,
-      });
-      contact.save();
-    })
-    .then(() => res.json({ success: true }))
-    .catch(errorHandler(res));
+router.patch("/edit/:id", idParamValidator(), upload.single("pfp"), (req, res) => {
+  if (req.file) {
+    const crop = JSON.parse(req.body.crop);
+
+    let compressor = sharp(req.file.path).rotate();
+    if (crop.width && crop.height && crop.left && crop.top) {
+      compressor = compressor.extract(crop);
+    }
+    compressor = compressor.resize(400, 400).png({
+      compressionLevel: 9,
+      adaptiveFiltering: true,
+      force: true,
+    });
+
+    Promise.all([
+      Contact.findById(req.params.id),
+      uploadFileStream(
+        compressor,
+        `pfp/${req.file.filename}-${Date.now()}-${Math.round(Math.random() * 1e9)}`
+      ),
+    ])
+      .then(([contact, result]) => {
+        const old = contact.profilePicture;
+
+        Object.assign(contact, {
+          name: req.body.updated_name ? req.body.updated_name : contact.name,
+          position: req.body.updated_position ? req.body.updated_position : contact.position,
+          organization: req.body.updated_org ? req.body.updated_org : contact.organization,
+          email: req.body.updated_email ? req.body.updated_email : contact.email,
+          phone: req.body.updated_phone ? req.body.updated_phone : contact.phone,
+          profilePicture: result.key,
+          profilePictureModified: new Date(),
+        });
+
+        return Promise.all([
+          contact.toJSON(),
+          contact.save(),
+          old ? deleteFileAWS(old) : null,
+          fs.unlink(req.file.path),
+        ]);
+      })
+      .then(() => res.json({ success: true }))
+      .catch(errorHandler(res));
+  } else {
+    Contact.findById(req.params.id)
+      .then((contact) => {
+        Object.assign(contact, {
+          name: req.body.updated_name ? req.body.updated_name : contact.name,
+          position: req.body.updated_position ? req.body.updated_position : contact.position,
+          organization: req.body.updated_org ? req.body.updated_org : contact.organization,
+          email: req.body.updated_email ? req.body.updated_email : contact.email,
+          phone: req.body.updated_phone ? req.body.updated_phone : contact.phone,
+        });
+        contact.save();
+      })
+      .then(() => res.json({ success: true }))
+      .catch(errorHandler(res));
+  }
 });
 
 /**
